Moved @auth_bp.route("/login" to auth_routes.py
 Changes to be committed:
	modified:   app.py
	modified:   routes/auth_routes.py
